Modify Random usage to allow reproducibility for the serial model.
There is a Random generator instance for every projection (discrete uniform)
and they all use the same seed. The seed is derived from the processor uptime.

For reproducibility, explicitly set rseed = 1. An original sim
spike train with unknown seed is in S2I10sp.orig . A sim with
rseed = 1 is in S2I10sp.std .
To compare spike trains use:
  nrngui spkplt.hoc -c 'p("S2I10sp.orig")'
The latter spike train will be overlaid in black on a ths standard spike
train in red.